mobile games
added homepage and 3 games, tic tac toe, number guess and reaction game. Made all pages in one single html doc and also added all css to styles.css and all scripts to scripts.js. This is to get a better overview of the project. Code is source formatted and validated.